full-blog/blog/requirements.txt
django
psycopg2-binary
djangorestframework
random-word
pyyaml
names
Faker
django-cors-headers

